Release 1.5.0: over-GUI denial messaging, perk-backlog drain, mod-perk purge, proper Apotheosis + gem gating
- Over-GUI denial-message system (OverlayNoticeGui/NoticeOverlayCP); skill-lock
  warning + spell-gating + Apotheosis affix gating now render over open screens
  (network protocol 5 -> 6).
- Implemented ~49 inert perks via PerkEffectsHandler (data-driven hooks).
- Removed all perks tied to Botania, Blood Magic, and Enigmatic Legacy (absent from
  the target pack): registrations, config, lang, textures, integration classes, lock
  providers, Botania build dep, and BOTANIA_RUNIC_SKILLS_INTEGRATION.md. Kept the
  Iron's Spellbooks blood-school perks, generic BLOOD_FURY, and the generic
  Enigmatic-flavored perks.
- Apotheosis: audited the 14 existing perks against the real 7.4.8 API (no drift),
  wired APOTHEOSIS_WISDOM via Placebo's GetEnchantmentLevelEvent, and added gem
  rarity gating (socketing gated by gem LootRarity -> Fortune level via
  ItemSocketingEvent.CanSocket; apothEnableGemRarityGating).
- Docs/version bumped to 1.5.0 across README, CurseForge, integration matrix, etc.

+++ src/main/java/com/otectus/runicskills/common/command/GlobalLimitCommand.java
@@ -16,7 +16,7 @@ public class GlobalLimitCommand {
         dispatcher.register((
                 Commands.literal("globallimit")
                         .requires((source) -> source.hasPermission(2))
-                        .then(Commands.argument("level", IntegerArgumentType.integer())
+                        .then(Commands.argument("level", IntegerArgumentType.integer(1))
                                 .executes(GlobalLimitCommand::execute)
                         )
         ));
